refactor(server): home accessibleSpaces/accessibleSpaceAlbums in scope module (slice 5)
Move the two scoping subqueries out of the upstream sync.repository.ts into the
fork-owned scope module (reducing fork footprint in an upstream file) and
re-export them from sync for its existing callers. No external importers; all 359
sync medium tests unchanged. Behavior identical.

Spec: data/sa-abstraction-spec-t8/report.md
